feat: add support for tables (#12)
As I was going through what's left to import the dev docs, noticed we
had a few tables over there, which this PR provides support for.

I also fixed a nasty bug when inserting top level elements, which was
due to incorrectly handling the cursor when inserting headings.

Finally, there's a curious hardcoded `200` value (with a comment of
course), which caused me to pull my hairs for a while: when you pass the
`extensions.GFM` for table support, the underlying code is also
attaching rendering functions for HTML rendering them, with a priority
of `500`.

While this is quite suprising, it's in the `goldmark` code and the
rationale for it is I suppose that it's just more convenient. I think we
could untangle that by using the lower level API for the configuring
options, but it's not very much worth the effort at this stage.