fix: tables can't be inserted w/o their children (#16)
Tables cannot be inserted without children. I think that this is only
for the rows, but rather than introducing a different detaching
strategy, chances that tables are too nested are pretty low as you can't
really do much that could create a block in there from Markdown.

So if we're adding blocks with the _walk_ strategy, if we see a table,
we append it with its children instead of walking through the children
and addding them one by one.